+++ lib/Views/UserInterface.jsx
@@ -1,5 +1,5 @@
 import {
-  Menu,
+  MenuLeft,
   Nav,
   ExperimentalMenu
 } from "terriajs/lib/ReactViews/StandardUserInterface/customizable/Groups";
@@ -14,40 +14,40 @@ import version from "../../version";
 
 import "./global.scss";
 
-function loadAugmentedVirtuality(callback) {
-  require.ensure(
-    "terriajs/lib/ReactViews/Map/Navigation/AugmentedVirtualityTool",
-    () => {
-      const AugmentedVirtualityTool = require("terriajs/lib/ReactViews/Map/Navigation/AugmentedVirtualityTool");
-      callback(AugmentedVirtualityTool);
-    },
-    "AugmentedVirtuality"
-  );
-}
+// function loadAugmentedVirtuality(callback) {
+//   require.ensure(
+//     "terriajs/lib/ReactViews/Map/Navigation/AugmentedVirtualityTool",
+//     () => {
+//       const AugmentedVirtualityTool = require("terriajs/lib/ReactViews/Map/Navigation/AugmentedVirtualityTool");
+//       callback(AugmentedVirtualityTool);
+//     },
+//     "AugmentedVirtuality"
+//   );
+// }
 
-function isBrowserSupportedAV() {
-  return /Android|iPhone|iPad/i.test(navigator.userAgent);
-}
+// function isBrowserSupportedAV() {
+//   return /Android|iPhone|iPad/i.test(navigator.userAgent);
+// }
 
 export default function UserInterface(props) {
   return (
     <StandardUserInterface {...props} version={version}>
-      <Menu>
-        <RelatedMaps viewState={props.viewState} />
+      <MenuLeft>
         <MenuItem caption="About" href="about.html" key="about-link" />
-      </Menu>
+        <RelatedMaps viewState={props.viewState} />
+      </MenuLeft>
       <Nav>
         <MeasureTool terria={props.viewState.terria} key="measure-tool" />
       </Nav>
       <ExperimentalMenu>
-        <If condition={isBrowserSupportedAV()}>
+        {/* <If condition={isBrowserSupportedAV()}>
           <SplitPoint
             loadComponent={loadAugmentedVirtuality}
             viewState={props.viewState}
             terria={props.viewState.terria}
             experimentalWarning={true}
           />
-        </If>
+        </If> */}
       </ExperimentalMenu>
     </StandardUserInterface>
   );